Added OneToMany and ManyToOne Relationships to User and Address Classes
Owner is the address class and the user variable in address class must be mapped to the list of addresses in user class
Builder excludes the addresses list initialisation so we should @Builder.Default for it consider.
The user object in Address should be excluded in ToString using @ToString.Exclude since the iterator goes into a stackoverflow error when printing the User.