Added ManyToMany Relationships to User and Tag Classes
Owner is the User class and the users variable in tag class must be mapped to the set of tags in user class
Builder excludes the addresses list initialisation so we should @Builder.Default for it consider.
The user object in Address should be excluded in ToString using @ToString.Exclude since the iterator goes into a stackoverflow error when printing the User.
Encapsulated the new tag creation and adding tags inside addTag Method in User class